move zipapp tools to tools/private
