+++ tests/tools/zipapp/BUILD.bazel
@@ -3,11 +3,11 @@ load("//python:py_test.bzl", "py_test")
 py_test(
     name = "zipper_test",
     srcs = ["zipper_test.py"],
-    deps = ["//tools/zipapp:zipper_lib"],
+    deps = ["//tools/private/zipapp:zipper_lib"],
 )
 
 py_test(
     name = "exe_zip_maker_test",
     srcs = ["exe_zip_maker_test.py"],
-    deps = ["//tools/zipapp:exe_zip_maker_lib"],
+    deps = ["//tools/private/zipapp:exe_zip_maker_lib"],
 )
